+++ Scripts/exec-swap-mission.sh
@@ -5,11 +5,11 @@
 #Description: Script that does the swap mission files to be executed automatically.
 #Descrição: Script que troca os arquivos da missão a ser executada de forma automática.
 
-#DIR_FILES=UAV-Mission-Creator/mission/kml/iros1
+DIR_FILES=UAV-Mission-Creator/mission/kml/iros1
 #DIR_FILES=UAV-Mission-Creator/mission/kml/iros2
 #DIR_FILES=UAV-Mission-Creator/mission/kml/iros3
 #DIR_FILES=UAV-Mission-Creator/mission/kml/pos-iros1
-DIR_FILES=UAV-Mission-Creator/mission/kml/icas
+#DIR_FILES=UAV-Mission-Creator/mission/kml/icas
 #DIR_FILES=UAV-Mission-Creator/mission/kml/campus2
 
 echo "=============swap file mission============="
@@ -22,6 +22,8 @@ cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/DE4s/map.sgl 2>/dev/null
 cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/GA4s/map.sgl 2>/dev/null
 cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/GH4s/map.sgl 2>/dev/null
 cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/MPGA4s/map.sgl 2>/dev/null
+cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/MS4s/map.sgl 2>/dev/null
+cp ../$DIR_FILES/map-full.sgl ../Modules-IFA/MILP4s/map.sgl 2>/dev/null
 
 echo "copy file map-nfz.sgl to MOSA"
 cp ../$DIR_FILES/map-nfz.sgl ../Modules-MOSA/HGA4m/. 2>/dev/null