+++ UAV-Generic/src/uav/generic/util/UtilGeom.java
@@ -4,12 +4,16 @@ import java.awt.Polygon;
 import java.awt.geom.Point2D;
 
 /**
- * Classe com métodos úteis sobre Geometria.
+ * Class with utils methods on Geometry.
  * @author Jesimar S. Arantes
  */
 public class UtilGeom {    
     
-    /*Calcula o centro X do poligono*/
+    /**
+     * Calc the center X of the polygon.
+     * @param poly the object polygon
+     * @return the center x of the polygon.
+     */
     public static double centerXPoly(Polygon poly){
         double sumX = 0;
         for (int i = 0; i < poly.npoints; i++) {
@@ -18,7 +22,11 @@ public class UtilGeom {
         return sumX/poly.npoints;
     }
     
-    /*Calcula o centro Y do poligono*/
+    /**
+     * Calc the center Y of the polygon.
+     * @param poly the object polygon
+     * @return the center y of the polygon.
+     */
     public static double centerYPoly(Polygon poly){
         double sumY = 0;
         for (int i = 0; i < poly.npoints; i++) {
@@ -27,7 +35,11 @@ public class UtilGeom {
         return sumY/poly.npoints;
     }
     
-    /*Calcula o centro X do poligono dado*/
+    /**
+     * Calc the center X of the polygon using the objetct Point2D.
+     * @param poly a vector of points that represent the polygon
+     * @return the center x of the polygon.
+     */
     public static double centerXPoly(Point2D poly[]){
         double sumX = 0;        
         for (Point2D item : poly) {
@@ -36,7 +48,11 @@ public class UtilGeom {
         return sumX/poly.length;
     }
     
-    /*Calcula o centro Y do poligono dado*/
+    /**
+     * Calc the center Y of the polygon using the objetct Point2D.
+     * @param poly a vector of points that represent the polygon
+     * @return the center y of the polygon.
+     */
     public static double centerYPoly(Point2D poly[]){
         double sumY = 0;        
         for (Point2D item : poly) {
@@ -45,14 +61,27 @@ public class UtilGeom {
         return sumY/poly.length;
     }        
     
-    /*Calcula a distância da posição (x, y) para o centro do poligono dado*/
+    /**
+     * Calc the distance of the position (x, y) to the center of polygon poly.
+     * @param poly the polygon to calc.
+     * @param x the position x.
+     * @param y the position y.
+     * @return the euclidian distante.
+     */
     public static double distPointToCenter(Point2D poly[], double x, double y){
         double x1 = centerXPoly(poly);
         double y1 = centerYPoly(poly);
         return Math.sqrt((x - x1)* (x - x1) + (y - y1) * (y - y1));
     }
       
-    /*Calcula a distancia euclidiana entre quatro pontos*/
+    /**
+     * Calc the euclidian distance amoung two points
+     * @param x1 coordinate x of first point.
+     * @param y1 coordinate y of first point.
+     * @param x2 coordinate x of second point.
+     * @param y2 coordinate y of second point.
+     * @return the euclidian distance.
+     */
     public static double distanceEuclidian(double x1, double y1, double x2, double y2){
         return Math.sqrt((x1 - x2)*(x1 - x2) + (y1 - y2) * (y1 - y2));
     }