+++ UAV-IFA/src/uav/ifa/module/path_replanner/GA_GA_4s.java
@@ -58,12 +58,12 @@ public class GA_GA_4s extends Replanner{
             String state = px + " " + py + " " + vel + " " + angle;
             String qtdWpt = configLocal.getQtdWaypoints();
             String delta = configLocal.getDelta();
-            UtilIO.copyFileMofifIFA(src, dst, state, 8, qtdWpt, 20, delta, 26);
+            UtilIO.copyFileModifiedIFA(src, dst, state, 8, qtdWpt, 20, delta, 26);
             
             File src_ga = new File(dirGA + "instance-base");
             File dst_ga = new File(dirGA + "instance");
             String time = configLocal.getTimeExec();
-            UtilIO.copyFileMofifIFA(src_ga, dst_ga, time, 117);
+            UtilIO.copyFileModifiedIFA(src_ga, dst_ga, time, 117);
             return true;
         } catch (FileNotFoundException ex) {
             StandardPrints.printMsgWarning("Warning [FileNotFoundException]: updateFileConfig()");