+++ UAV-IFA/src/uav/ifa/module/path_replanner/MPGA4s.java
@@ -56,12 +56,12 @@ public class MPGA4s extends Replanner{
             String state = px + " " + py + " " + vel + " " + angle;
             String qtdWpt = configLocal.getQtdWaypoints();
             String delta = configLocal.getDelta();
-            UtilIO.copyFileMofifIFA(src, dst, state, 8, qtdWpt, 20, delta, 26);
+            UtilIO.copyFileModifiedIFA(src, dst, state, 8, qtdWpt, 20, delta, 26);
             
             File src_mpga = new File(dir + "instance-base");
             File dst_mpga = new File(dir + "instance");
             String time = configLocal.getTimeExec();
-            UtilIO.copyFileMofifIFA(src_mpga, dst_mpga, time, 117);
+            UtilIO.copyFileModifiedIFA(src_mpga, dst_mpga, time, 117);
             return true;
         } catch (FileNotFoundException ex) {
             StandardPrints.printMsgWarning("Warning [FileNotFoundException]: updateFileConfig()");