+++ UAV-MOSA/src/uav/mosa/module/path_planner/CCQSP4m.java
@@ -2,7 +2,6 @@ package uav.mosa.module.path_planner;
 
 import java.io.File;
 import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.io.PrintStream;
 import java.util.Scanner;
 import lib.color.StandardPrints;
@@ -13,56 +12,37 @@ import uav.generic.struct.Waypoint;
 import uav.generic.util.UtilGeo;
 import uav.generic.util.UtilIO;
 import uav.generic.hardware.aircraft.Drone;
-import uav.generic.hardware.aircraft.FixedWing;
 
 /**
- * Classe que modela o planejador de rotas HGA4m. 
+ * Classe que modela o planejador de rotas CCQSP4m. 
  * @author Jesimar S. Arantes
  */
-public class HGA4m extends Planner{
+public class CCQSP4m extends Planner{
     
     /**
      * Class constructor
      * @param drone instance of the aircraft
      * @param waypointsMission waypoints of the mission
      */
-    public HGA4m(Drone drone, Mission3D waypointsMission) {
+    public CCQSP4m(Drone drone, Mission3D waypointsMission) {
         super(drone, waypointsMission);
     }   
     
-    @Override
-    public boolean execMission(int i) {
-        boolean itIsOkUpdate = updateFileConfig(i);
-        boolean itIsOkpathAB = definePathAB(i);        
+    public boolean execMission() {
+        boolean itIsOkUpdate = updateFileConfig();     
         boolean itIsOkExec   = execMethod();
-        boolean itIsOkRoute  = createFileFinalRoute(i);
-        boolean itIsOkParse  = parseRoute3DtoGeo(i);
-        return itIsOkUpdate && itIsOkpathAB && itIsOkExec && itIsOkRoute && itIsOkParse;
+        boolean itIsOkParse  = parseRoute3DtoGeo();
+        return itIsOkUpdate && itIsOkExec && itIsOkParse;
     }
     
-    @Override
-    public boolean updateFileConfig(int i) {
+    public boolean updateFileConfig() {
         try {
-            double px1 = waypointsMission.getPosition3D(i).getX();
-            double py1 = waypointsMission.getPosition3D(i).getY();
-            double px2 = waypointsMission.getPosition3D(i+1).getX();
-            double py2 = waypointsMission.getPosition3D(i+1).getY();
-            double dx = px2 - px1;
-            double dy = py2 - py1;
-            //distancia entre os pontos com uma margem de seguranca
-            double dist = Math.sqrt(dx*dx+dy*dy)*2;
-            
-            File src_ga = new File(dir + "ga-config-base");
-            File dst_ga = new File(dir + "ga-config");
-            String time = configLocal.getTimeExec(i);
-            String timeH = String.format("%d", (int)(dist));
-            //usando metade dos waypoints DeltaT=2
-            String qtdWpt = String.format("%d", (int)(dist/2));
-            String delta = configLocal.getDelta();
-            String maxVel = configLocal.getMaxVelocity();
-            String maxCtrl = configLocal.getMaxControl();
-            UtilIO.copyFileMofifMOSA(src_ga, dst_ga, time, 207, delta, 304,
-                    qtdWpt, 425, timeH, 426, maxVel, 427, maxCtrl, 428);
+            File src_instance = new File(dir + "instance-base");
+            File dst_instance = new File(dir + "instance");
+            String delta = configLocal.getDeltaCCQSP4m();
+            String qtdWpt = configLocal.getWaypointsCCQSP4m();
+            UtilIO.copyFileModifiedMOSA(src_instance, dst_instance, 
+                    delta, 189, qtdWpt, 298, qtdWpt, 299);
             return true;
         } catch (FileNotFoundException ex) {
             StandardPrints.printMsgWarning("Warning [FileNotFoundException]: updateFileConfig()");
@@ -70,103 +50,21 @@ public class HGA4m extends Planner{
         }
     }
     
-    private double vx1 = 0.0;
-    private double vy1 = 0.0;
-    private boolean definePathAB(int i) {
+    public boolean parseRoute3DtoGeo(){
         try {
-            double px1 = waypointsMission.getPosition3D(i).getX();
-            double py1 = waypointsMission.getPosition3D(i).getY();
-            double px2 = waypointsMission.getPosition3D(i+1).getX();
-            double py2 = waypointsMission.getPosition3D(i+1).getY();
-            double vx2 = 0;
-            double vy2 = 0;
-            if (i < waypointsMission.size() - 2){
-                double px3 = waypointsMission.getPosition3D(i+2).getX();
-                double py3 = waypointsMission.getPosition3D(i+2).getY();
-                double dx = px3 - px1;
-                double dy = py3 - py1;
-                double norm = Math.sqrt(dx*dx+dy*dy);
-                double vc = drone.getSpeedCruize();
-                vx2 = dx * vc/norm;
-                vy2 = dy * vc/norm;
-            }   
-            
-            PrintStream print = new PrintStream(new File(dir + "mission-config.sgl"));
-            print.println("----------- start state (px py vx vy) -----------");
-            if (drone instanceof FixedWing){
-                print.println(px1 + "," + py1 + "," + vx1 + "," + vy1);
-            } else {
-                print.println(px1 + "," + py1 + ",0.0,0.0");
-            }
-            print.println("--------------- end point (px py)---------------");
-            if (drone instanceof FixedWing){
-                print.println(px2 + "," + py2 + "," + vx2 + "," + vy2);
-            } else {
-                print.println(px2 + "," + py2 + ",0.0,0.0");
-            }
-            print.println();
-            print.println("<TrueName>");
-            print.println("Config2D-2.sgl");
-            print.close();
-            vx1 = vx2;
-            vy1 = vy2;
-            return true;
-        } catch (FileNotFoundException ex) {
-            StandardPrints.printMsgWarning("Warning [FileNotFoundException]: definePathAB()");
-            return false;
-        } 
-    }  
-    
-    private boolean createFileFinalRoute(int i) {
-        try {
-            File src = new File(dir + "output-simulation.log");
-            File dst = new File(dir + "out-sim"+i+".log");
-            UtilIO.copyFile(src, dst);
-            File route = new File(dir + "route3D"+i+".txt");
-            
-            Scanner sc = new Scanner(dst);
-            PrintStream print = new PrintStream(route);
-            boolean test1 = false;
-            boolean test2 = false;
-            while (sc.hasNextLine()) {
-                String line = sc.nextLine();
-                if (!test1 && !test2){
-                    if (line.equals("------------------------ [ states ] ------------------------")){
-                        test1 = true;
-                    }
-                }else if (test1 && !test2){
-                    if (!line.equals("------------------------ [ controls ] ------------------------")){
-                        print.println(line);
-                    }else{
-                        test2 = true;
-                    }
-                }
-            }
-            sc.close();
-            print.close();
-            return true;
-        } catch (IOException ex) {
-            StandardPrints.printMsgWarning("Warning [IOException]: createFileFinalRoute()");
-            return false;
-        }
-    }
-    
-    @Override
-    public boolean parseRoute3DtoGeo(int i){
-        try {
-            String nameFileRoute3D =  "route3D"  + i + ".txt";
-            String nameFileRouteGeo = "routeGeo" + i + ".txt";
+            String nameFileRoute3D =  "output.txt";
+            String nameFileRouteGeo = "routeGeo.txt";
             PointGeo pGeoBase = UtilGeo.getPointGeo(configGlobal.getDirFiles() + 
                     configGlobal.getFileGeoBase());
             File fileRouteGeo = new File(dir + nameFileRouteGeo);
             PrintStream printGeo = new PrintStream(fileRouteGeo);
             Scanner readRoute3D = new Scanner(new File(dir + nameFileRoute3D));
             int countLines = 0;
+            readRoute3D.nextInt();
             while(readRoute3D.hasNext()){
                 double x = readRoute3D.nextDouble();
                 double y = readRoute3D.nextDouble();
-                readRoute3D.nextDouble();
-                readRoute3D.nextDouble();
+                readRoute3D.nextInt();
                 double h = configGlobal.getAltRelMission();            
                 printGeo.println(UtilGeo.parseToGeo(pGeoBase, x, y, h, ";"));
                 mission3D.addPosition3D(new Position3D(x, y, h));
@@ -192,6 +90,7 @@ public class HGA4m extends Planner{
     public void clearLogs() {
         UtilIO.deleteFile(new File(dir), ".log");
         UtilIO.deleteFile(new File(dir), ".png");
+        UtilIO.deleteFile(new File(dir), ".err");
         new File(dir + "log_error.txt").delete(); 
     }